Issue-39: Created MainScreenVM.kt and use it as parent for TvsVM, PersonsVM, MoviesVM. Refactored data classes in data package.
