--- app/src/main/java/com/kirchhoff/movies/data/DiscoverMoviesResponse.kt
@@ -1,8 +1,0 @@
-package com.kirchhoff.movies.data
-
-data class DiscoverMoviesResponse(
-    val page: Int,
-    val results: List<Movie>,
-    val total_results: Int,
-    val total_pages: Int
-)
