--- app/src/main/java/com/kirchhoff/movies/data/DiscoverTvsResponse.kt
@@ -1,8 +1,0 @@
-package com.kirchhoff.movies.data
-
-data class DiscoverTvsResponse(
-    val page: Int,
-    val results: List<Tv>,
-    val total_results: Int,
-    val total_pages: Int
-)
