--- app/src/main/java/com/kirchhoff/movies/data/Movie.kt
@@ -1,8 +1,0 @@
-package com.kirchhoff.movies.data
-
-data class Movie(
-    val id: Int,
-    val title: String,
-    val poster_path: String?,
-    val backdrop_path: String?
-)
