--- app/src/main/java/com/kirchhoff/movies/data/Person.kt
@@ -1,7 +1,0 @@
-package com.kirchhoff.movies.data
-
-data class Person(
-    val id: Int,
-    val name: String,
-    val profile_path: String?
-)
