--- app/src/main/java/com/kirchhoff/movies/data/PersonsResponse.kt
@@ -1,8 +1,0 @@
-package com.kirchhoff.movies.data
-
-data class PersonsResponse(
-    val page: Int,
-    val results: List<Person>,
-    val total_results: Int,
-    val total_pages: Int
-)
