--- app/src/main/java/com/kirchhoff/movies/data/Tv.kt
@@ -1,7 +1,0 @@
-package com.kirchhoff.movies.data
-
-data class Tv(
-    val id: Int,
-    val poster_path: String?,
-    val name: String
-)
