+++ app/src/main/java/com/kirchhoff/movies/repository/person/IPersonsRepository.kt
@@ -1,6 +1,6 @@
 package com.kirchhoff.movies.repository.person
 
-import com.kirchhoff.movies.data.PersonsResponse
+import com.kirchhoff.movies.data.responses.PersonsResponse
 import com.kirchhoff.movies.repository.Result
 
 interface IPersonsRepository {
