+++ app/src/main/java/com/kirchhoff/movies/repository/person/PersonsRepository.kt
@@ -1,6 +1,6 @@
 package com.kirchhoff.movies.repository.person
 
-import com.kirchhoff.movies.data.PersonsResponse
+import com.kirchhoff.movies.data.responses.PersonsResponse
 import com.kirchhoff.movies.network.services.PersonService
 import com.kirchhoff.movies.repository.BaseRepository
 import com.kirchhoff.movies.repository.Result
